Added null checks in AlgorithmController. Removed redundant double imageId checks in AnnotationController.

+++ src/main/java/com/manuscript/rest/api/AlgorithmController.java
@@ -7,7 +7,9 @@ import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Objects;
 import java.util.UUID;
+import java.util.stream.Stream;
 
 import static com.manuscript.rest.common.Constants.RESOURCE_ALGORITHM;
 
@@ -20,34 +22,47 @@ public class AlgorithmController {
 
     @PostMapping("/runAlgorithm")
     public void runAlgorithm(@RequestBody AlgorithmRequest algorithmRequest) {
+        checkNotNull(algorithmRequest);
         algorithmService.run(algorithmRequest);
     }
 
     @PostMapping("/uploadAlgorithm")
     public ResponseEntity<AlgorithmResponse> uploadAlgorithm(@RequestBody AlgorithmRequest algorithmRequest) {
+        checkNotNull(algorithmRequest);
         AlgorithmResponse result = algorithmService.create(algorithmRequest);
         return ResponseEntity.ok(result);
     }
 
     @PutMapping("/updateAlgorithm")
     public ResponseEntity<AlgorithmResponse> updateAlgorithm(@RequestBody AlgorithmRequest algorithmRequest) {
+        checkNotNull(algorithmRequest);
         AlgorithmResponse result = algorithmService.update(algorithmRequest);
         return ResponseEntity.ok(result);
     }
 
-    @GetMapping("/getAlgorithm/{algorithmId}")
-    public ResponseEntity<AlgorithmResponse> getAlgorithm(@PathVariable UUID algorithmId) {
+    @GetMapping("/getAlgorithm/{algorithmId}/{uid}")
+    public ResponseEntity<AlgorithmResponse> getAlgorithm(@PathVariable UUID algorithmId, @PathVariable String uid) {
+        if(algorithmId == null || uid == null)
+            throw new IllegalArgumentException("Invalid algorithm user ID.");
         AlgorithmResponse result = algorithmService.getById(algorithmId);
         return ResponseEntity.ok(result);
     }
 
     @DeleteMapping("/deleteAlgorithm/{algorithmId}/{uid}")
     public void deleteAlgorithm(@PathVariable UUID algorithmId, @PathVariable String uid) {
+        if(algorithmId == null || uid == null)
+            throw new IllegalArgumentException("Invalid algorithm or user ID.");
         algorithmService.delete(algorithmId, uid);
     }
 
     @DeleteMapping("/deleteAllByUserIdAlgorithm")
     public void deleteAllByUserIdAlgorithm(@RequestBody UUID userId) {
         algorithmService.deleteAllByUserId(userId);
+    }
+
+    private void checkNotNull(AlgorithmRequest algorithmRequest) {
+        if(Stream.of(algorithmRequest.getId(), algorithmRequest.getUid(), algorithmRequest.getImageId(),
+                        algorithmRequest.getUrl()).anyMatch(Objects::isNull))
+            throw new IllegalArgumentException("Algorithm request's fields must not be null.");
     }
 }
